# docs/swagger.yaml
basePath: /api/v1
definitions:
  database.ApplicationStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - ApplicationStatusPending
    - ApplicationStatusApproved
    - ApplicationStatusRejected
...
  database.PartnerApplication:
    properties:
      audienceSize:
        type: string
      companyName:
        type: string
      createdAt:
        type: string
      id:
        type: string
      promotionPlan:
        type: string
      rejectionReason:
        type: string
      reviewedAt:
        type: string
      reviewedBy:
        type: string
      socialMedia:
        additionalProperties: true
        type: object
      status:
        $ref: '#/definitions/database.ApplicationStatus'
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/database.User'
      userId:
        type: string
      website:
        type: string
    type: object
...
  database.User:
    properties:
      binanceEmail:
        type: string
      canTrade:
        type: boolean
      canWithdraw:
        type: boolean
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isWhitelistEnabled:
        type: boolean
      kycLevel:
        type: integer
      lastName:
        type: string
      metaData:
        additionalProperties: true
        type: object
      status:
        type: integer
      subAccountId:
        type: string
      updatedAt:
        type: string
      vipLevel:
        type: integer
    type: object
...
  structs.ApplicationListParams:
    properties:
      current:
        type: integer
      export:
        type: boolean
      limit:
        description: alias for PageSize
        type: integer
      page:
        description: alias for Current
        type: integer
      pageSize:
        type: integer
      query:
        type: string
      sortDate:
        $ref: '#/definitions/structs.SortDate'
      status:
        type: string
    type: object
  structs.ApplicationReviewParams:
    properties:
      rejectionReason:
        type: string
    type: object
  structs.ChartParams:
    properties:
      endDate:
        type: string
      groupBy:
        description: '"day", "week", "month"'
        type: string
      startDate:
        type: string
    type: object
...
  structs.DashboardSummaryParams:
    properties:
      endDate:
        type: string
      startDate:
        type: string
    type: object
  structs.ErrorResponse:
    properties:
      body:
        type: string
      error_msg:
        type: string
      status_code:
        type: integer
    type: object
  structs.PaginationInput:
    properties:
      current:
        type: integer
      export:
        type: boolean
      limit:
        description: alias for PageSize
        type: integer
      page:
        description: alias for Current
        type: integer
      pageSize:
        type: integer
      sortDate:
        $ref: '#/definitions/structs.SortDate'
    type: object
  structs.PaginationResponse:
    properties:
      items: {}
      total:
        type: integer
    type: object
...
  structs.ResponseBody:
    properties:
      body: {}
      message:
        type: string
    type: object
  structs.SortDate:
    properties:
      end_day:
        type: string
      start_day:
        type: string
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: |-
    Partner program backend — public, partner, admin, system endpoints.
    Authentication is split into three layers:
    • BearerAuth (PartnerAuth/AdminAuth): Cognito ID token in `Authorization: Bearer …`.
    • InternalKey: `X-Internal-Key` header — only the xmeta-monorepo server should send this.
  title: xmeta Partner API
  version: "1.0"
paths:
  /admin/partner/analytics/commission-trend:
    post:
      consumes:
      - application/json
      description: Returns time-bucketed commission totals across all partners
      parameters:
      - description: Chart range filter
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.ChartParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ResponseBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Commission trend chart
      tags:
      - Admin Analytics
  /admin/partner/analytics/referral-funnel:
    post:
      consumes:
      - application/json
      description: Returns conversion-funnel metrics from click to commissioned trade
      parameters:
      - description: Date range filter
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.DashboardSummaryParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ResponseBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Referral funnel metrics
      tags:
      - Admin Analytics
  /admin/partner/analytics/summary:
    post:
      consumes:
      - application/json
      description: Returns top-level KPIs for the admin analytics dashboard
      parameters:
      - description: Date range filter
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.DashboardSummaryParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ResponseBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin analytics summary
      tags:
      - Admin Analytics
  /admin/partner/analytics/top-partners:
    post:
      consumes:
      - application/json
      description: Returns the highest-earning partners ranked by commission volume
      parameters:
      - description: Pagination input
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.PaginationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/structs.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Top partners leaderboard
      tags:
      - Admin Analytics
  /admin/partner/applications/{id}/approve:
    post:
      consumes:
      - application/json
      description: Marks an application approved and provisions the partner account
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/database.PartnerApplication'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve partner application
      tags:
      - Admin Applications
  /admin/partner/applications/{id}/reject:
    post:
      consumes:
      - application/json
      description: Marks an application rejected with a reason
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.ApplicationReviewParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/database.PartnerApplication'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject partner application
      tags:
      - Admin Applications
  /admin/partner/applications/detail/{id}:
    get:
      consumes:
      - application/json
      description: Returns one partner application with full submission data
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/database.PartnerApplication'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get application detail
      tags:
      - Admin Applications
  /admin/partner/applications/list:
    post:
      consumes:
      - application/json
      description: Returns a paginated list of partner applications for admin review
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.ApplicationListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/structs.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List partner applications
      tags:
      - Admin Applications
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: '"Bearer {Cognito ID Token}" format'
    in: header
    name: Authorization
    type: apiKey
  InternalKey:
    description: Server-to-server API key shared with xmeta-monorepo
    in: header
    name: X-Internal-Key
    type: apiKey
swagger: "2.0"